pyproject.toml
[project]
name = "ab_engine"
version = "0.1.2"
description = "Library of standard modules"
authors = [{name = "Andrew Bulatov", email = "BulatovAndrew@gmail.com"}]
license = "MIT"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "json5>=0.12.0",
    "psycopg>=3.2.9",
    "pyaml>=25.7.0",
    "toml>=0.10.2",
]
[project.urls]
Repository = "https://github.com/a-bulatov/ab_engine"
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
